refactor: move user to constant
