+++ src/utils/constants/index.ts
@@ -1,5 +1,6 @@
 import Time from './time';
+import User from './user';
 
 const IS_PUBLIC_KEY = 'isPublic';
 
-export { Time, IS_PUBLIC_KEY };
+export { Time, User, IS_PUBLIC_KEY };
