+++ src/utils/constants/user.ts
@@ -1,0 +1,10 @@
+const user = {
+  email: 'test@gmail.com',
+  password: 'Hello@1',
+};
+
+const activateUser = {
+  email: 'nguyenkhoa438@gmail.com',
+};
+
+export default { user, activateUser };
